rootpy/interactive: Fix rootwait crashes once and for all. Add canvas_events.py
rootwait crashes were fixed by stopping the GUI thread and restarting
it when we are done.

In addition, also made it possible to optionally close canvases by
middle clicking them or pressing escape when they have focus (ROOT requires
the mouse to be over the pad area for this to work). This is provided in
the canvas_events module.